requirements.txt
# Core ML and NLP dependencies
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
spacy>=3.7.0
spacy-transformers>=1.3.0
scikit-learn>=1.3.0
numpy>=1.24.0,<2.0.0
nltk>=3.8.0

# Vector storage backends
faiss-cpu>=1.7.4
chromadb>=0.4.0

# LLM integration
huggingface-hub>=0.17.0  # Required for Hugging Face Inference API

# Web framework
streamlit>=1.25.0
faster-whisper>=1.0.0

# Data handling and visualization
pandas>=2.0.0
plotly>=5.0.0

# PDF generation
reportlab>=4.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0  # Required for OpenStreetMap integration
pywhatkit>=5.3  # Required for WhatsApp alerts